keyword-matcher-python/requirements.txt
htmllaundry~=2.2
requests~=2.28.1
nats-py==2.1.4
lxml~=4.6.3
pydantic~=1.9.1
